Site for 0.4.0

+++ documentation/com/io7m/jcamera/examples/jogl/ExampleTimer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2014 <code@io7m.com> http://io7m.com
+ * Copyright © 2016 <code@io7m.com> http://io7m.com
  * 
  * Permission to use, copy, modify, and/or distribute this software for any
  * purpose with or without fee is hereby granted, provided that the above
@@ -59,7 +59,7 @@ public final class ExampleTimer
       {
         while (true) {
           try {
-            Thread.sleep(Integer.MAX_VALUE);
+            Thread.sleep((long) Integer.MAX_VALUE);
           } catch (final InterruptedException ex) {
             // Nothing
           }
